+++ tests/functional/L0_example_linear_cupy/test.py
@@ -50,6 +50,7 @@ def main():
     args = parser.parse_args()
 
     logging.basicConfig(level=logging.DEBUG, format=DEFAULT_LOG_FORMAT)
+    logging.captureWarnings(True)
 
     docker_image_with_name = METADATA["image_name"].format(TEST_CONTAINER_VERSION=get_current_container_version())
     verify_docker_image_in_readme_same_as_tested("examples/linear_cupy/README.md", docker_image_with_name)