+++ tests/functional/L0_response_cache/test.py
@@ -50,6 +50,7 @@ def main():
 
     log_level = logging.DEBUG if args.verbose else logging.INFO
     logging.basicConfig(level=log_level, format=DEFAULT_LOG_FORMAT)
+    logging.captureWarnings(True)
     LOGGER.debug(f"CLI args: {args}")
 
     class _InferFuncWrapper: