+++ tests/functional/L0_run_in_thread/server.py
@@ -90,6 +90,7 @@ def main():
 
     log_level = logging.DEBUG if args.verbose else logging.INFO
     logging.basicConfig(level=log_level, format=DEFAULT_LOG_FORMAT)
+    logging.captureWarnings(True)
     LOGGER.debug(f"CLI args: {args}")
 
     random.seed(args.seed)