+++ tests/functional/L1_futures_client_stress/test.py
@@ -69,6 +69,7 @@ def main():
 
     log_level = logging.DEBUG if args.verbose else logging.INFO
     logging.basicConfig(level=log_level, format=DEFAULT_LOG_FORMAT)
+    logging.captureWarnings(True)
 
     futures_stress_test(
         test_time_s=args.test_time_s,