+++ tests/functional/L1_hf_nlp_distilbert/test.py
@@ -76,6 +76,7 @@ def main():
 
     log_level = logging.DEBUG if args.verbose else logging.INFO
     logging.basicConfig(level=log_level, format=DEFAULT_LOG_FORMAT)
+    logging.captureWarnings(True)
 
     huggingface_distilbert(
         test_time_s=args.test_time_s,