+++ tests/functional/L2_hf_nlp_distilbert/test.py
@@ -16,6 +16,7 @@
 Runs longer (10h) inference session over NLP model
 """
 import argparse
+import logging
 import random
 
 METADATA = {
@@ -67,6 +68,8 @@ def main():
     args = parser.parse_args()
 
     random.seed(args.seed)
+
+    logging.captureWarnings(True)
 
     with TestMonitoringContext(args):
         huggingface_distilbert(